move class definition to header file
